requirements.txt
Django>=4.1.6
mysqlclient>=2.1
numpy>=1.24.2
matplotlib>=3.6.3
mysql-connector-python>=8.0.32
djangorestframework~=3.14.0
Pillow~=9.4.0